Separate inactive user filtering tests into dedicated files
- Create user_inactive_filtering_test.go for API4 tests
- Create user_inactive_filtering_e2e_test.go for mmctl E2E tests
- Remove inactive filtering tests from original test files
- Maintain test functionality while improving organization
